Move BPM Task Status enum to constatns module
