+++ constants/src/main/java/org/jboss/pnc/enums/BPMTaskStatus.java
@@ -15,12 +15,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.restmodel.bpm;
+package org.jboss.pnc.enums;
 
 /**
  * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
  */
-public enum BpmTaskStatus {
+public enum BPMTaskStatus {
     STARTING,
     STARTED,
     COMPLETED;
